Create geom_*_trace layers

Create geom_*_trace layers

create_trace_layers(
  mapping,
  data,
  stat,
  geom,
  position,
  show.legend,
  inherit.aes,
  params,
  trace_position,
  background_params,
  trans_fn,
  allow_bottom = TRUE
)

Arguments

mapping
Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).

stat
The statistical transformation to use on the data for this layer, as a string.

geom
The geometric object to use to display the data.

position
Position adjustment, either as a string, or the result of a call to a position adjustment function.

show.legend
logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes
If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

params
Additional parameters to pass to the geom and stat.

trace_position
Specifies which groups of data points should be outlined. Can be 'all', 'bottom', or a predicate to use for filtering data. If 'all', the default, every group plotted will be outlined, if 'bottom', only the bottom most layer of points will be outlined. A subset of data points can be outlined by passing a predicate. This must evaluate to `TRUE` or `FALSE` within the context of the input data.

background_params
Named list specifying aesthetic parameters to use for background points when a predicate is passed to `trace_position`.

trans_fn
Function to use for transforming data when predicate is passed to trace_position. Must accept three arguments: dat, data to transform; ex, expression to use for transforming data; inv, should the expression be negated, this allows the inverse data points to be transformed.

allow_bottom
Should 'bottom' be allowed as an argument for trace_position?