Inspect the call stack on enter

Inspect the call stack on enter

Usage

ggtraceback(method, ...)

Arguments

method

A function or a ggproto method. The ggproto method may be specified using any of the following forms:

ggproto$method

namespace::ggproto$method

namespace:::ggproto$method

...

Arguments passed on to rlang::trace_back

top

The first frame environment to be included in the backtrace. This becomes the top of the backtrace tree and represents the oldest call in the backtrace.

This is needed in particular when you call trace_back() indirectly or from a larger context, for example in tests or inside an RMarkdown document where you don't want all of the knitr evaluation mechanisms to appear in the backtrace.

If not supplied, the rlang_trace_top_env global option is consulted. This makes it possible to trim the embedding context for all backtraces created while the option is set. If knitr is in progress, the default value for this option is knitr::knit_global() so that the knitr context is trimmed out of backtraces.

bottom

The last frame environment to be included in the backtrace. This becomes the rightmost leaf of the backtrace tree and represents the youngest call in the backtrace.

Set this when you would like to capture a backtrace without the capture context.

Can also be an integer that will be passed to caller_env().

Details

Calls rlang::trace_back() upon first entering the function or method. Note that ggtraceback() only ever prints the back trace once because it calls gguntrace() after the first time it's triggered. For more complex inspections of the call stack, ggedit() is recommended.

Note

The output of rlang::trace_back() is not logged to the tracedumps (last_ggtrace() and global_ggtrace())