Create geom_*_trace layers

Create geom_*_trace layers

create_trace_layers(
  mapping,
  data,
  stat,
  geom,
  position,
  show.legend,
  inherit.aes,
  params,
  trace_position,
  background_color,
  allow_bottom = TRUE
)

Arguments

mapping
Set of aesthetic mappings created by aes() to use for generating layers.

data
The data to be displayed in these layers.

stat
Statistical transformation to use for these layers.

geom
The geometric object to use to display the data.

position
Position adjustment.

show.legend
Should these layers be included in the legend?

inherit.aes
If FALSE, overrides the default aesthetics.

params
Additional parameters to pass to the geom and stat.

trace_position
Specifies which groups of data points should be outlined. Can be 'all', 'bottom', or a predicate to use for filtering data. If 'all', the default, every group plotted will be outlined, if 'bottom', only the bottom most layer of points will be outlined. A subset of data points can be outlined by passing a predicate. This must evaluate to `TRUE` or `FALSE` within the context of the input data.

background_color
Color to use for background points when a predicate is passed to `trace_position`. If NULL, the original fill color will be used.

allow_bottom
Should 'bottom' be allowed as an argument for trace_position?